fix reset InternetGw issue for Windows
