+++ wireguard/wireguard.go
@@ -136,31 +136,6 @@ func GetPeer(ifaceName, peerPubKey string) (wgtypes.Peer, error) {
 	return wgtypes.Peer{}, fmt.Errorf("peer not found")
 }
 
-// getDefaultGatewayIpFromRouteList - an internal function to get the default gateway ip from route list string
-func getDefaultGatewayIpFromRouteList(output string) string {
-
-	var rList []string
-	if strings.Contains(output, "\r") {
-		rList = strings.Split(output, "\r")
-	} else if strings.Contains(output, "\n") {
-		rList = strings.Split(output, "\n")
-	}
-
-	var rLine string
-	for _, l := range rList {
-		if strings.Contains(l, "0.0.0.0") {
-			rLine = l
-			if strings.Contains(l, ncutils.GetInterfaceName()) {
-				break
-			}
-		}
-	}
-
-	rLineList := strings.Fields(rLine)
-
-	return strings.TrimSpace(rLineList[len(rLineList)-1])
-}
-
 // GetOriginalDefaulGw - fetches system's original default gw
 func GetOriginalDefaulGw() (link int, gwIP net.IP, err error) {
 	link = config.Netclient().OriginalDefaultGatewayIfLink
